requirements.txt
requests
customtkinter
openpyxl
Pillow